+++ src/adapters/akamai/index.entry.js
@@ -4,18 +4,19 @@
  * @description Main entry point for the Akamai EdgeWorker
  */
 // CDN specific imports
-import AkamaiAdapter from './cdn-adapters/akamai/akamaiAdapter';
-import AkamaiKVInterface from './cdn-adapters/akamai/akamaiKVInterface';
+import AkamaiAdapter from './akamaiAdapter';
+import AkamaiKVInterface from './akamaiKVInterface';
 
 // Application specific imports
-import CoreLogic from './coreLogic'; // Assume this is your application logic module
-import OptimizelyProvider from './_optimizely_/optimizelyProvider';
-import defaultSettings from './_config_/defaultSettings';
-import * as optlyHelper from './_helpers_/optimizelyHelper';
-import { getAbstractionHelper } from './_helpers_/abstractionHelper';
-import Logger from './_helpers_/logger';
-import EventListeners from './_event_listeners_/eventListeners';
-import handleRequest from './_api_/apiRouter';
+import CoreLogic from '../../core/providers/coreLogic'; // Assume this is your application logic module
+import OptimizelyProvider from '../../core/providers/optimizelyProvider';
+import defaultSettings from '../../utils/config/defaultSettings';
+import * as optlyHelper from '../../utils/optimizelyHelper';
+import { getAbstractionHelper } from '../../utils/abstractionHelper';
+import Logger from '../../utils/logging/logger';
+import EventListeners from '../../core/providers/events/eventListeners';
+import handleRequest from '../../core/api/apiRouter';
+
 //
 let abstractionHelper, logger;
 // Define the request, environment, and context objects after initializing the AbstractionHelper
@@ -140,7 +141,7 @@ export async function onClientRequest(request) {
 	const datafileOperation = pathName === '/v1/datafile';
 
 	// Check if the request is for the config operation
-	const configOperation = pathName === '/v1/config';
+	const configOperation = pathName === '/v1/config";
 
 	// Check if the sdkKey is provided in the request headers
 	let sdkKey = _abstractRequest.getHeader(defaultSettings.sdkKeyHeader);
