+++ src/adapters/vercel/vercelAdapter.js
@@ -1,0 +1,4 @@
+import * as optlyHelper from '../../utils/optimizelyHelper';
+import * as cookieDefaultOptions from '../../utils/config/cookieOptions';
+import defaultSettings from '../../utils/config/defaultSettings';
+import EventListeners from '../../core/providers/events/eventListeners';
